app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
email-validator==2.2.0
dnspython==2.6.1
jinja2==3.1.4
passlib==1.7.4
PyJWT==2.8.0